requirements.txt
PyPDF2
python-dotenv
streamlit
openai
faiss-cpu
altair
tiktoken
langchain-community
langchain-openai
langchain